requirements.txt
ccxt
colorama
python-dotenv
eth-account
hyperliquid-python-sdk
matplotlib
pandas
pandas_ta
yfinance
cryptography
